Disentangle SpecimenApiController from study-main
