+++ study/api-src/org/labkey/api/study/StudyUtils.java
@@ -1,11 +1,18 @@
 package org.labkey.api.study;
 
+import org.jetbrains.annotations.NotNull;
 import org.labkey.api.data.SQLFragment;
 import org.labkey.api.data.dialect.SqlDialect;
+import org.labkey.api.specimen.Vial;
 import org.labkey.api.util.DateUtil;
 
 import java.util.Calendar;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.Date;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 // Utility methods that are shared between study and specimen
 public class StudyUtils
@@ -27,4 +34,47 @@ public class StudyUtils
         sql.append("(").append(dialect.getDatePart(Calendar.DAY_OF_MONTH, dateColumnName)).append(")");
         return sql;
     }
+
+    @NotNull
+    private static <T> Set<T> intersect(@NotNull Set<T> left, @NotNull Set<T> right)
+    {
+        Set<T> intersection = new HashSet<>();
+        for (T item : left)
+        {
+            if (right.contains(item))
+                intersection.add(item);
+        }
+        return intersection;
+    }
+
+    @NotNull
+    public static Collection<Integer> getPreferredProvidingLocations(Collection<List<Vial>> specimensBySample)
+    {
+        Set<Integer> locationIntersection = null;
+
+        for (List<Vial> vials : specimensBySample)
+        {
+            Set<Integer> currentLocations = new HashSet<>();
+            for (Vial vial : vials)
+            {
+                if (vial.getCurrentLocation() != null)
+                    currentLocations.add(vial.getCurrentLocation());
+            }
+            if (locationIntersection == null)
+                locationIntersection = currentLocations;
+            else
+            {
+                locationIntersection = intersect(locationIntersection, currentLocations);
+                if (locationIntersection.isEmpty())
+                    return locationIntersection;
+            }
+        }
+
+        if (null != locationIntersection)
+            return locationIntersection;
+
+        return Collections.emptySet();
+    }
+
+
 }
